// src/shared/interfaces/DoctorPayload.ts
export interface DoctorPayload {
	staffId: number;
	thumbnail: string;
	fullName: string;
	specialtyId: number;
	specialtyName: string;
	clinicId: number;
	clinicName: string;
	department: string;
	yearsOfExperience: number;
	education: string;
	introduce: string;
	member: string;
	expert: string;
}